rOstluft long format (rolf)

Data format for storing air quality data. A Serie is the combination of measurement site, interval, parameter and unit.

The data is chunked in year of measurement, site and interval. Base64 encoding is used to avoid invalid characters

Usage

format_rolf(tz = "Etc/GMT-1")

Arguments

tz

time zone used for chunking. Default Etc/GMT-1

Value

R6 class object of format_rolf

Fields

index

index for the data (in most cases the time columns)

serie_columns

combination of columns for a serie

chunk_columns

data columns for chunking

chunk_calc

formulas for calculating new columns for chunking

unique_columns

= index + serie_columns

content_columns

= serie_columns + names(chunk_calc)

tz

time zone used for chunking. Same time different chunk: 2018-01-01T00:00:00+0100 = 2018, 2017-12-31T23:00:00+0000 = 2017

Methods

$sort(data)

$merge(new_data, old_data)

$chunk_name(chunk_data) returns the chunkname based on the first row of the supplied data

$get_chunk_names(interval, site, year) returns encoded chunknames for all combinations of the supplied vectors

$encode_chunk_name(interval, site, year) returns encoded chunkname

$decode_chunk_name(chunk_name) returns decoded chunk_name in tibble as columns (chunk_name, interval, site, year)

$destroy(confirmation) removes all files under path from the file system if "DELETE" is supplied as confirmation

TODO

Update Documentation is outdated