Returns data of the Tecson meteo station placed at Tiefenbrunnen and Mythenquai water police station.

Uses the Tecdottir API from Stefan Oderholz.

get_seepolizei(start, end, site = c("tiefenbrunnen", "mythenquai"))

Arguments

start
The start date after which the measurements should be returned. As string in format yyyy-mm-dd or as POSIXct, POSIXlt or Date

end
The end date before which the measurements should be returned. As string in format yyyy-mm-dd or as POSIXct, POSIXlt or Date

site
The site of which the values should be returned. Valid values are "mythenquai" and "tiefenbrunnen"

Value

tibble in rolf format

Examples

get_seepolizei(as.Date("2020-03-01"), lubridate::dmy("02.03.2020"))

#> # A tibble: 2,890 x 6
#>    starttime           site          parameter       interval unit  value
#>    <dttm>              <fct>         <fct>           <fct>    <fct> <dbl>
#>  1 2020-03-01 00:00:00 tiefenbrunnen air_temperature min10    °C      6.5
#>  2 2020-03-01 00:10:00 tiefenbrunnen air_temperature min10    °C      6.5
#>  3 2020-03-01 00:20:00 tiefenbrunnen air_temperature min10    °C      6.4
#>  4 2020-03-01 00:30:00 tiefenbrunnen air_temperature min10    °C      6.1
#>  5 2020-03-01 00:40:00 tiefenbrunnen air_temperature min10    °C      6  
#>  6 2020-03-01 00:50:00 tiefenbrunnen air_temperature min10    °C      6  
#>  7 2020-03-01 01:00:00 tiefenbrunnen air_temperature min10    °C      6  
#>  8 2020-03-01 01:10:00 tiefenbrunnen air_temperature min10    °C      5.9
#>  9 2020-03-01 01:20:00 tiefenbrunnen air_temperature min10    °C      6  
#> 10 2020-03-01 01:30:00 tiefenbrunnen air_temperature min10    °C      6.1
#> # ... with 2,880 more rows

data <- get_seepolizei("2020-03-20", "2020-03-22", "mythenquai")

# normalize data with meta informations
meta_data <- readRDS(rOstluft.data::f("meta_seepolizei.rds"))

# rename station
data <- meta_apply(
  data = data,
  meta = meta_data,
  data_src = "site",
  data_dest = "site",
  meta_key = "site_short",
  meta_val = "site",
  mode = "strict"
)

# rename parameters to Ostluft Conventions
data <- meta_apply(
  data = data,
  meta = meta_data,
  data_src = "parameter",
  data_dest = "parameter",
  meta_key = "parameter_original",
  meta_val = "parameter",
  mode = "strict"
)

# show normalized data
data

#> # A tibble: 3,897 x 6
#>    starttime           site          parameter interval unit  value
#>    <dttm>              <fct>         <fct>     <fct>    <fct> <dbl>
#>  1 2020-03-20 00:00:00 ZH_Mythenquai T         min10    °C     10  
#>  2 2020-03-20 00:10:00 ZH_Mythenquai T         min10    °C      9.9
#>  3 2020-03-20 00:20:00 ZH_Mythenquai T         min10    °C     10  
#>  4 2020-03-20 00:30:00 ZH_Mythenquai T         min10    °C      9.9
#>  5 2020-03-20 00:40:00 ZH_Mythenquai T         min10    °C      9.8
#>  6 2020-03-20 00:50:00 ZH_Mythenquai T         min10    °C      9.8
#>  7 2020-03-20 01:00:00 ZH_Mythenquai T         min10    °C      9.7
#>  8 2020-03-20 01:10:00 ZH_Mythenquai T         min10    °C      9.6
#>  9 2020-03-20 01:20:00 ZH_Mythenquai T         min10    °C      9.5
#> 10 2020-03-20 01:30:00 ZH_Mythenquai T         min10    °C      9.3
#> # ... with 3,887 more rows