Reads an AIRMO Export in CSV Webexport Format

Reads the data from a AIRMO Webexport CSV file and returns it as a data frame in rOstluft long format

read_airmo_webexport(
  fn,
  encoding = "latin1",
  tz = "Etc/GMT-1",
  time_shift = NULL,
  na.rm = TRUE
)

Arguments

fn
CSV file name, can be relative to current directory or absolut

encoding
CSV file encoding. Default "latin1"

tz
time zone of date field. Be carefull Etc/GMT + == -. Default "Etc/GMT-1"

time_shift
a lubridate period to add to the time. Default NULL

na.rm
remove na (empty) values. Default TRUE

Value

data frame in rOstluft long format

See also

Examples

input <- rOstluft.data::f("ol_nox_covid19_2020.csv")
read_airmo_webexport(input)

#> # A tibble: 60,916 x 6
#>    starttime           site         parameter interval unit  value
#>    <dttm>              <fct>        <fct>     <fct>    <fct> <dbl>
#>  1 2020-01-01 00:00:00 Opf_Balsberg NOx       min30    ppb    20.0
#>  2 2020-01-01 00:30:00 Opf_Balsberg NOx       min30    ppb    21.5
#>  3 2020-01-01 01:00:00 Opf_Balsberg NOx       min30    ppb    26.1
#>  4 2020-01-01 01:30:00 Opf_Balsberg NOx       min30    ppb    21.5
#>  5 2020-01-01 02:00:00 Opf_Balsberg NOx       min30    ppb    27.3
#>  6 2020-01-01 02:30:00 Opf_Balsberg NOx       min30    ppb    19.9
#>  7 2020-01-01 03:00:00 Opf_Balsberg NOx       min30    ppb    17.7
#>  8 2020-01-01 03:30:00 Opf_Balsberg NOx       min30    ppb    17.9
#>  9 2020-01-01 04:00:00 Opf_Balsberg NOx       min30    ppb    19.7
#> 10 2020-01-01 04:30:00 Opf_Balsberg NOx       min30    ppb    23.2
#> # ... with 60,906 more rows