Derive metadata (parameters, units etc) from an IAC/ETHZ meteo station file

Reads metadata from meteo station measurement files (10 minute time resolution) by the institute for atmosphere and climety at ETH Zürich. Currently, two stations in Zürich are under long-term operation: CHN-Gebäude & Hönggerberg (see http://www.iac.ethz.ch/the-institute/weather-stations.html)

read_ethz_iac_min10_parameters(x, site = NULL)

Arguments

x
Either a path to a file, a connection, or literal data. see readr::read_delim()

site
Character string specifying the site of the meteo station. Needs to be exactly like in file name, e.g. "-Hberg" for files from Hönggerberg station. For site 'CHN Gebäude' the default of site is NULL (due to respective file naming)

Value

tibble in rOstluft long format structure