Read a file from Nabel

Reads Exports from National Air Pollution Monitoring Network (NABEL) The National Air Pollution Monitoring Network (NABEL) measures air pollution at 16 locations in Switzerland. The stations are distributed throughout the country and monitor pollution at typical locations (e.g. city-centre streets, residential areas, rural stations). The monitoring network has commenced operations in stages since 1979 and is operated by the Federal office for the environment and Empa

This function reads the parameter and unit information from the header. The interval is auto detected if possible. In Addition the time information are in end time. The time is converted to start time and the time zone defined trough the argument tz. The argument time_shift provides a way to manuelly shift the time series. In this case no automatically shifting is applied. The provided values is directly added to information in the file.

Usage

read_nabel_txt(
  fn,
  encoding = "latin1",
  tz = "Etc/GMT-1",
  interval = NULL,
  time_shift = NULL,
  na.rm = TRUE
)

Arguments

fn

path to input file

encoding

encoding of the data file. Default = "latin1"

tz

of the data. Default "Etc/GMT-1"

interval

optional interval of the data. Use if auto detect fails. Default NULL. If used it is necessary to define time_shift manuelly. lubridate::period(0) can be used for no shifting

time_shift

a lubridate period to add to the time. Default NULL

na.rm

remove na values. Default TRUE

Value

tibble in rolf format