Reads a file from the SwissMetNet of MeteoSwiss

SwissMetNet, the automatic monitoring network of MeteoSwiss the Federal Office for Meteorology and Climatology, comprises about 160 automatic monitoring stations. These stations deliver a multitude of current data on the weather and climate in Switzerland every ten minutes. The monitoring network is supplemented by automatic precipitation stations.

read_smn(fn, tz = "Etc/GMT-1", encoding = "UTF-8",
  time_format = NULL, interval = NULL, na.rm = TRUE)

Arguments

fn
path to input file

tz
of the output data. Default "Etc/GMT-1"

encoding
encoding of the data file. Default = "UTF-8"

time_format
optional time_format. Use if auto detect fails. Default NULL

interval
optional interval of the data. Use if auto detect fails. Default NULL

na.rm
remove na values. Default TRUE

Value

tibble in rOstluft long format structure

Examples

input <- system.file("extdata", "smn.txt", package = "rOstluft.data", mustWork = TRUE)
read_smn(input)

#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 01:00:00 CHU   dkl010z0  min10    NA    216  
#>  2 2018-01-01 01:10:00 CHU   dkl010z0  min10    NA    209  
#>  3 2018-01-01 01:20:00 CHU   dkl010z0  min10    NA    216  
#>  4 2018-01-01 01:30:00 CHU   dkl010z0  min10    NA    227  
#>  5 2018-01-01 01:40:00 CHU   dkl010z0  min10    NA    212  
#>  6 2018-01-01 01:50:00 CHU   dkl010z0  min10    NA    200  
#>  7 2018-01-01 01:00:00 CHU   fkl010z0  min10    NA      1.7
#>  8 2018-01-01 01:10:00 CHU   fkl010z0  min10    NA      1.9
#>  9 2018-01-01 01:20:00 CHU   fkl010z0  min10    NA      2  
#> 10 2018-01-01 01:30:00 CHU   fkl010z0  min10    NA      2.6
#> # ... with 44 more rows