Resampling data

Aggregate data by different time periods. Following this simple steps:

split data in series

pad data serie (needed for calculation of capture threshold)

group serie by new interval with lubridate::floor_date()

apply statistical method or user provides function (user can provide list per parameter)

combine resampled series

resample(data, new_interval, statistic = "mean", data.thresh = NULL,
  start.date = NULL, end.date = NULL, drop.last = FALSE,
  percentile = 95)

resample_series(serie, new_interval, statistic = "mean",
  data.thresh = NULL, start.date = NULL, end.date = NULL,
  drop.last = FALSE, percentile = 95)

Arguments

data
A tibble in rOstluft long format

new_interval
New interval. Must be longer than actual interval (not checked)

statistic
Statistical method to apply when aggregating the data; default is the mean. Can be one of “mean”, “max”, “min”, “median”, “frequency”, “sd”, “percentile”. Note that “sd” is the standard deviation, “frequency” is the number (frequency) of valid records in the period and “data.cap” is the percentage data capture. “percentile” is the percentile level (%) between 0-100, which can be set using the “percentile” argument. Or a function with one argument expecting a vector. Or a list with parameter as name and the statistical method as value (function or name of method)

data.thresh
minimum data capture threshold in % to use

start.date
optional start date for padding. Default min date in series floored to the new interval

end.date
optional end date for padding. Default max date in series ceiled to the new interval

drop.last
optional drop the last added time point by padding. Default False, true if no end.date provided and max date != ceiled max date.

percentile
The percentile level in % used when statistic = "percentile". The default is 95.

serie
a tibble in rOstluft long format containing exactly one serie

Value

tibble with resampled data

TODO Wind

Need a solution for wind vectoring ... this isn't as simple. probably need group all wind and non wind parameters. Handle wind like now and do something special with the wind vars. One possible to group:

dg <- dplyr::group_by(data, isWind = .data$parameter %in% c("WD", "WVv", "WVs"))
tidyr::nest()
A tibble: 2 x 2
 isWind data
 <lgl>  <list>
 1 FALSE  <tibble [208,344 x 6]>
 2 TRUE   <tibble [35,037 x 6]>

This was the easy part :P. now join the wind data, calc vector mean and return.