s3 storage

A s3 storage with flexible file format (default rds). The data format defines the data chunks per file. The data is cached locally. This local cache can be used as local storage. For performance enhancement it is recommended after acquiring all needed data from s3 to use the cache as local storage. See storage_local_rds() for more Informations.

storage_s3_rds(
  name,
  format,
  bucket,
  prefix = NULL,
  region = NULL,
  read.only = TRUE
)

Arguments

name
name of the store

format
data format of the store

bucket
name in aws s3

prefix
in aws s3

region
aws region

read.only
read only store. disables put, upload

Value

R6 class object of storage_s3

Fields

name

name of the store

format

data format of the store

bucket

s3 bucket containing the store

region

aws region of the bucket

prefix

of the s3 object keys. see Object Key and Metadata

path

local root of the store

data_path

local root of all chunks

data_s3

s3 root key of all chunks

content_path

local path to the rds file containing statistics of store content

content_s3

s3 object key to the rds file containing statistics of store content

columns_path

local path to the rds file containing the exact column types of the store content

columns_s3

s3 object key to the rds file containing the exact column types of the store content

meta_path

local root of all meta data files

meta_s3

s3 root key of all meta data files

read.only

flag for read.only usage of store. Default TRUE

ext

file extension for chunks. Default "rds"

read_function

function(file) for reading chunks from disk. Default base::readRDS()

write_function

function(object, file) for writing chunks to disk. Default base::saveRDS()

Methods

$get(filter=NULL, ...) get data from the store. The name of the arguments depend on the format. The filter argument is applied to each chunk.

$download(...) downloads data from s3 to the local cache. The dots arguments are used to filter the output of list_chunks(). Only the matching rows will be downloaded.

$put(data) puts the data into the store. Stops if store is read only

$upload() uploads content, meta data and all new and changed chunks to the s3 storage. For big additions to the store the recommend way is to use a local storage to modify the cache and use this function to apply the change.

$get_content() returns a tibble with the amount of data points per chunk per series

$list_chunks() get list of all chunks in s3 and local

$get_meta(key=NULL) get meta data. If key is omitted returns all the content of all files in a named list of tibbles, with the file name without extension as name. If key is supplied as argument only the list contains only the specified key.

$put_meta(...) puts meta data into the store. the name of the argument is used as file name and the value as data.

$get_local_storage() returns a storage to work with the cached data like a local storage

$destroy(confirmation) removes all files under path from the file system if "DELETE" is supplied as confirmation

Authentication

See the documentation of aws.signature for ways to provide the necessary informations. The simplest way is to use environment variables defined in a .Renviron file in the root directory of a RStudio Project:

AWS_ACCESS_KEY_ID = "xxxxxxxxxxxxxxxxxx"
AWS_SECRET_ACCESS_KEY = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
AWS_DEFAULT_REGION = "eu-central-1"

Examples

## init store, creates directory if necessary
format <- rOstluft::format_rolf()
store <- rOstluft::storage_s3_rds("s3_example", format, "rostluft", prefix = "aqmet")
#> store s3_example initialized under 'C:\Users\tom\AppData\Local/rOstluft/s3_example'

## get all data min30 for 2011 and 2012
store$get(site = "Zch_Stampfenbachstrasse", interval = "min30", year = 2011:2012)
#> # A tibble: 523,318 x 6
#>    starttime           site                    parameter interval unit   value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct>  <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.19 
#>  2 2011-01-01 00:00:00 Zch_Stampfenbachstrasse CO        min30    ppm    0.525
#>  3 2011-01-01 00:00:00 Zch_Stampfenbachstrasse Hr        min30    %Hr   91.9  
#>  4 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO2       min30    ppb   22.0  
#>  5 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO        min30    ppb   21.0  
#>  6 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NOx       min30    ppb   43.0  
#>  7 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10 
#>  8 2011-01-01 00:00:00 Zch_Stampfenbachstrasse PM10      min30    µg/m3 84.0  
#>  9 2011-01-01 00:00:00 Zch_Stampfenbachstrasse RainDur   min30    min    0    
#> 10 2011-01-01 00:00:00 Zch_Stampfenbachstrasse SO2       min30    ppb    1.90 
#> # ... with 523,308 more rows

## get only data for O3
store$get(year = 2011:2012, site = "Zch_Stampfenbachstrasse", interval = "min30",
          filter = parameter == "O3")
#> # A tibble: 69,992 x 6
#>    starttime           site                    parameter interval unit  value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct> <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.19
#>  2 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10
#>  3 2011-01-01 00:30:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.10
#>  4 2011-01-01 00:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.05
#>  5 2011-01-01 01:00:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.10
#>  6 2011-01-01 01:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.05
#>  7 2011-01-01 01:30:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.00
#>  8 2011-01-01 01:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.00
#>  9 2011-01-01 02:00:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.04
#> 10 2011-01-01 02:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.02
#> # ... with 69,982 more rows

## get NOx data from multiple stations
store$get(site = c("Zch_Stampfenbachstrasse", "Zch_Rosengartenstrasse"), interval = "min30",
          year = 2014, filter = parameter %in% c("NOx", "NO", "NO2"))
#> # A tibble: 104,547 x 6
#>    starttime           site                   parameter interval unit  value
#>    <dttm>              <fct>                  <fct>     <fct>    <fct> <dbl>
#>  1 2014-01-01 00:00:00 Zch_Rosengartenstrasse NO2       min30    ppb    18.6
#>  2 2014-01-01 00:00:00 Zch_Rosengartenstrasse NO        min30    ppb    42.8
#>  3 2014-01-01 00:00:00 Zch_Rosengartenstrasse NOx       min30    ppb    61.4
#>  4 2014-01-01 00:30:00 Zch_Rosengartenstrasse NO2       min30    ppb    21.3
#>  5 2014-01-01 00:30:00 Zch_Rosengartenstrasse NO        min30    ppb    55.9
#>  6 2014-01-01 00:30:00 Zch_Rosengartenstrasse NOx       min30    ppb    77.2
#>  7 2014-01-01 01:00:00 Zch_Rosengartenstrasse NO2       min30    ppb    28.2
#>  8 2014-01-01 01:00:00 Zch_Rosengartenstrasse NO        min30    ppb    88.1
#>  9 2014-01-01 01:00:00 Zch_Rosengartenstrasse NOx       min30    ppb   116. 
#> 10 2014-01-01 01:30:00 Zch_Rosengartenstrasse NO2       min30    ppb    39.8
#> # ... with 104,537 more rows

## get n data points grouped by intervall, station, parameter, year in the store
store$get_content()
#> # A tibble: 193,549 x 6
#>     year interval site          parameter unit      n
#>    <dbl> <fct>    <fct>         <fct>     <fct> <int>
#>  1  2020 y1       Wgi_Weiertal  NH3       µg/m3     1
#>  2  2020 y1       WeF_Weid      NH3       µg/m3     1
#>  3  2020 y1       Wdi_Waldhof   NH3       µg/m3     1
#>  4  2020 y1       Tan_NABEL     NH3       µg/m3     1
#>  5  2020 y1       SHn_Ebnet     NH3       µg/m3     1
#>  6  2020 y1       Rap_Hochhaus  NH3       µg/m3     1
#>  7  2020 y1       Oeg_St.Anton  NH3       µg/m3     1
#>  8  2020 y1       LaR_Grabewis  NH3       µg/m3     1
#>  9  2020 y1       Gon_Gontenbad NH3       µg/m3     1
#> 10  2020 y1       Fra_Sand      NH3       µg/m3     1
#> # ... with 193,539 more rows

## get list of all chunks, show only local files
dplyr::filter(store$list_chunks(), !is.na(local.path))
#> # A tibble: 4 x 11
#>   chunk_name    interval site   year  s3.key    s3.lastmodified s3.etag  s3.size
#>   <fs::path>    <chr>    <chr>  <chr> <chr>     <chr>           <chr>    <fs::b>
#> 1 min30/bWluMz~ min30    Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 4dce887~ 808.31K
#> 2 min30/bWluMz~ min30    Zch_S~ 2014  aqmet/da~ 2021-04-18T08:~ 7539d5f~   1.22M
#> 3 min30/bWluMz~ min30    Zch_S~ 2011  aqmet/da~ 2021-04-18T08:~ e65eff0~   1.21M
#> 4 min30/bWluMz~ min30    Zch_S~ 2012  aqmet/da~ 2021-04-18T08:~ 6e51bfa~   1.22M
#> # ... with 3 more variables: local.path <fs::path>,
#> #   local.modification_time <dttm>, local.size <fs::bytes>

## download all data for site Zch_Rosengartenstrasse before 2016
store$download(site == "Zch_Rosengartenstrasse", year < 2016)
#> list()

## now there should be 2 more local files
dplyr::filter(store$list_chunks(), !is.na(local.path))
#> # A tibble: 20 x 11
#>    chunk_name   interval site   year  s3.key    s3.lastmodified s3.etag  s3.size
#>    <fs::path>   <chr>    <chr>  <chr> <chr>     <chr>           <chr>    <fs::b>
#>  1 d1/ZDHCu1pj~ d1       Zch_R~ 2013  aqmet/da~ 2021-04-18T08:~ 7d432cd~  19.84K
#>  2 d1/ZDHCu1pj~ d1       Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 19630b9~  76.76K
#>  3 d1/ZDHCu1pj~ d1       Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 4909442~  86.04K
#>  4 h1/aDHCu1pj~ h1       Zch_R~ 2013  aqmet/da~ 2021-04-18T08:~ 7a8a700~ 112.38K
#>  5 h1/aDHCu1pj~ h1       Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ a289ca4~ 447.54K
#>  6 h1/aDHCu1pj~ h1       Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 2382520~ 567.64K
#>  7 h8gl/aDhnbM~ h8gl     Zch_R~ 2013  aqmet/da~ 2021-04-18T08:~ c0c2b8c~  23.56K
#>  8 h8gl/aDhnbM~ h8gl     Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 4f559bf~  93.14K
#>  9 h8gl/aDhnbM~ h8gl     Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 1db43d3~   92.2K
#> 10 m1/bTHCu1pj~ m1       Zch_R~ 2013  aqmet/da~ 2021-04-18T08:~ 8c9f389~   2.08K
#> 11 m1/bTHCu1pj~ m1       Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 175b6ac~   5.64K
#> 12 m1/bTHCu1pj~ m1       Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 4798271~   6.02K
#> 13 min30/bWluM~ min30    Zch_R~ 2013  aqmet/da~ 2021-04-18T08:~ 891111f~ 203.55K
#> 14 min30/bWluM~ min30    Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 4dce887~ 808.31K
#> 15 min30/bWluM~ min30    Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 026f2ae~ 924.68K
#> 16 min30/bWluM~ min30    Zch_S~ 2014  aqmet/da~ 2021-04-18T08:~ 7539d5f~   1.22M
#> 17 min30/bWluM~ min30    Zch_S~ 2011  aqmet/da~ 2021-04-18T08:~ e65eff0~   1.21M
#> 18 min30/bWluM~ min30    Zch_S~ 2012  aqmet/da~ 2021-04-18T08:~ 6e51bfa~   1.22M
#> 19 y1/eTHCu1pj~ y1       Zch_R~ 2014  aqmet/da~ 2021-04-18T08:~ 4f48153~   1.49K
#> 20 y1/eTHCu1pj~ y1       Zch_R~ 2015  aqmet/da~ 2021-04-18T08:~ 3c127a6~   1.48K
#> # ... with 3 more variables: local.path <fs::path>,
#> #   local.modification_time <dttm>, local.size <fs::bytes>

## get all meta data
store$get_meta()
#> $all
#> # A tibble: 2,057 x 12
#>    site    site_short site_long        x      y  masl parameter_origi~ parameter
#>    <chr>   <chr>      <chr>        <dbl>  <dbl> <dbl> <chr>            <chr>    
#>  1 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 Hr               Hr       
#>  2 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 NO               NO       
#>  3 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 NO2              NO2      
#>  4 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 NOx              NOx      
#>  5 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 O3               O3       
#>  6 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 Ox               Ox       
#>  7 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 p                p        
#>  8 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 PM10             PM10     
#>  9 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 SO2              SO2      
#> 10 Arb_St~ ARBS       Arbon Stad~ 2.75e6 1.26e6   407 StrGlo           StrGlo   
#> # ... with 2,047 more rows, and 4 more variables: unit <chr>, interval <chr>,
#> #   timezone_original <chr>, source <chr>
#> 
#> $ecmwf
#> # A tibble: 96 x 14
#>    site_short site   site_long   lon   lat masl  parameter_origi~ unit  interval
#>    <chr>      <chr>  <lgl>     <dbl> <dbl> <lgl> <chr>            <chr> <chr>   
#>  1 ZH_EraI    Züric~ NA         8.25  47.2 NA    t2m              K     h1      
#>  2 ZH_EraI    Züric~ NA         8.25  47.2 NA    swvl1            m3/m3 h1      
#>  3 ZH_EraI    Züric~ NA         8.25  47.2 NA    sp               hPa   h1      
#>  4 ZH_EraI    Züric~ NA         8.25  47.2 NA    u10              m/s   h1      
#>  5 ZH_EraI    Züric~ NA         8.25  47.2 NA    v10              m/s   h1      
#>  6 ZH_EraI    Züric~ NA         8.25  47.2 NA    skt              K     h1      
#>  7 ZH_EraI    Züric~ NA         8.25  47.2 NA    ishf             W/m2  h1      
#>  8 ZH_EraI    Züric~ NA         8.25  47.2 NA    tp               m     h1      
#>  9 ZH_EraI    Züric~ NA         8.25  47.2 NA    tco3             kg/m2 h1      
#> 10 ZH_EraI    Züric~ NA         8.25  47.2 NA    sund             s     h1      
#> # ... with 86 more rows, and 5 more variables: timezone_original <chr>,
#> #   source <chr>, x <dbl>, y <dbl>, parameter <chr>
#> 
#> $ethz
#> # A tibble: 28 x 14
#>    parameter_origin~ unit  site_short site_long site    lon   lat  masl interval
#>    <chr>             <chr> <chr>      <lgl>     <chr> <dbl> <dbl> <dbl> <chr>   
#>  1 p_air             hPa   ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  2 T_air             °C    ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  3 T_dew             °C    ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  4 RH                %     ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  5 illum             log1~ ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  6 rain              mm    ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  7 wspd_vec          m/s   ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  8 wspd_scl          m/s   ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#>  9 wdir              °     ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#> 10 wspd_pk           m/s   ETHZ_CHN   NA        ETHZ~  8.55  47.4   510 min10   
#> # ... with 18 more rows, and 5 more variables: timezone_original <chr>,
#> #   source <chr>, x <dbl>, y <dbl>, parameter <chr>
#> 
#> $hysplit
#> # A tibble: 2 x 12
#>   site_short  site  site_long lon   lat   masl  parameter_origi~ parameter unit 
#>   <chr>       <chr> <chr>     <lgl> <lgl> <lgl> <chr>            <chr>     <lgl>
#> 1 ZH-Kaserne~ ZH_K~ Zürich_K~ NA    NA    NA    hysplit_cluster  hysplit_~ NA   
#> 2 SG-BP-hysp~ SG_B~ St. Gall~ NA    NA    NA    hysplit_cluster  hysplit_~ NA   
#> # ... with 3 more variables: interval <chr>, timezone_original <chr>,
#> #   source <chr>
#> 
#> $hysplit_cluster_trajectories
#> # A tibble: 1,164 x 7
#> # Groups:   cluster [6]
#>    cluster hour.inc default            lat   lon date                site       
#>      <dbl>    <dbl> <fct>            <dbl> <dbl> <dttm>              <fct>      
#>  1       1      -96 01 January 1980~  48.6 -13.3 2000-07-20 06:49:23 SG_Blumenb~
#>  2       1      -95 01 January 1980~  48.5 -13.0 2000-07-20 06:49:23 SG_Blumenb~
#>  3       1      -94 01 January 1980~  48.5 -12.7 2000-07-20 06:49:23 SG_Blumenb~
#>  4       1      -93 01 January 1980~  48.5 -12.4 2000-07-20 06:49:23 SG_Blumenb~
#>  5       1      -92 01 January 1980~  48.4 -12.1 2000-07-20 06:49:23 SG_Blumenb~
#>  6       1      -91 01 January 1980~  48.4 -11.8 2000-07-20 06:49:23 SG_Blumenb~
#>  7       1      -90 01 January 1980~  48.3 -11.5 2000-07-20 06:49:23 SG_Blumenb~
#>  8       1      -89 01 January 1980~  48.3 -11.2 2000-07-20 06:49:23 SG_Blumenb~
#>  9       1      -88 01 January 1980~  48.2 -10.9 2000-07-20 06:49:23 SG_Blumenb~
#> 10       1      -87 01 January 1980~  48.2 -10.6 2000-07-20 06:49:23 SG_Blumenb~
#> # ... with 1,154 more rows
#> 
#> $meteoschweiz
#> # A tibble: 265 x 15
#>    site_short site    Länge Breite       x      y  masl parameter_origi~ source 
#>    <chr>      <chr>   <chr> <chr>    <dbl>  <dbl> <dbl> <chr>            <chr>  
#>  1 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 tre200s0         MeteoS~
#>  2 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 ure200s0         MeteoS~
#>  3 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 prestas0         MeteoS~
#>  4 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 prestas0         MeteoS~
#>  5 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 fkl010z1         MeteoS~
#>  6 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 fkl010z0         MeteoS~
#>  7 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 dkl010z0         MeteoS~
#>  8 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 sre000z0         MeteoS~
#>  9 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 sre000z0         MeteoS~
#> 10 TAE        Aadorf~ 8°54' 47°29' 2710517 1.26e6   539 gre000z0         MeteoS~
#> # ... with 255 more rows, and 6 more variables: unit <chr>, Beschreibung <chr>,
#> #   timezone_original <chr>, site_long <lgl>, interval <chr>, parameter <chr>
#> 
#> $meteotest
#> # A tibble: 2 x 12
#>   site_short site_long site   lon   lat    masl parameter_origi~ parameter unit 
#>   <chr>      <lgl>     <chr>  <lgl> <lgl> <dbl> <chr>            <chr>     <chr>
#> 1 SMA_23m    NA        Züric~ NA    NA      581 ta1tows0         T         °C   
#> 2 ZUEberg    NA        Züric~ NA    NA      729 ta1tows0         T         °C   
#> # ... with 3 more variables: interval <chr>, timezone_original <chr>,
#> #   source <chr>
#> 
#> $nabel
#> # A tibble: 406 x 31
#>     msID site   site_short msNameKurz  site_long  msKT  msOrt msOrtsteil msNABEL
#>    <int> <chr>  <chr>      <chr>       <chr>      <chr> <chr> <chr>        <int>
#>  1  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  2  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  3  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  4  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  5  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  6  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  7  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  8  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#>  9  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#> 10  1084 Züric~ NA         Zürich-Kas~ Zürich-Ka~ ZH    Züri~ NABEL            1
#> # ... with 396 more rows, and 22 more variables: scDTV <chr>, scDTVJahr <chr>,
#> #   scLKWProzent <int>, scStrassenabstand <int>, scEinwohner <int>,
#> #   scEinwohnerJahr <chr>, scVerkehrslage <chr>, scSiedlungsgroesse <chr>,
#> #   scLagerelativSiedlung <chr>, scIDBStandorttyp <chr>,
#> #   scIDBStationstyp <chr>, scIDBBebauung <chr>, scIDBVerkehr <chr>, x <int>,
#> #   y <int>, masl <int>, msmmBeginn <dttm>, msmmEnde <dttm>, source <chr>,
#> #   timezone_original <chr>, parameter_original <chr>, parameter <chr>
#> 
#> $ostluft
#> # A tibble: 1,560 x 31
#>     msID site    site_short msNameKurz  site_long msKT  msOrt msOrtsteil msNABEL
#>    <int> <chr>   <chr>      <chr>       <chr>     <chr> <chr> <chr>        <int>
#>  1     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  2     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  3     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  4     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  5     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  6     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  7     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  8     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  9     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#> 10     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#> # ... with 1,550 more rows, and 22 more variables: scDTV <chr>,
#> #   scDTVJahr <chr>, scLKWProzent <int>, scStrassenabstand <int>,
#> #   scEinwohner <int>, scEinwohnerJahr <chr>, scVerkehrslage <chr>,
#> #   scSiedlungsgroesse <chr>, scLagerelativSiedlung <chr>,
#> #   scIDBStandorttyp <chr>, scIDBStationstyp <chr>, scIDBBebauung <chr>,
#> #   scIDBVerkehr <chr>, x <int>, y <int>, masl <int>, parameter <chr>,
#> #   msmmBeginn <dttm>, msmmEnde <dttm>, source <chr>, ...
#> 
#> $seepolizei
#> # A tibble: 30 x 14
#>    parameter_origin~ unit  site_short site_long site    lon   lat  masl interval
#>    <chr>             <chr> <chr>      <lgl>     <chr> <dbl> <dbl> <dbl> <chr>   
#>  1 air_temperature   °C    mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  2 water_temperature °C    mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  3 wind_gust_max_10~ m/s   mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  4 wind_speed_avg_1~ m/s   mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  5 wind_force_avg_1~ bft   mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  6 wind_direction    °     mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  7 windchill         °C    mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  8 barometric_press~ hPa   mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#>  9 precipitation     mm    mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#> 10 dew_point         °C    mythenquai NA        ZH_M~  8.54  47.4   407 min10   
#> # ... with 20 more rows, and 5 more variables: timezone_original <chr>,
#> #   source <chr>, x <dbl>, y <dbl>, parameter <chr>
#> 

## or a specific meta file
store$get_meta("ostluft")
#> $ostluft
#> # A tibble: 1,560 x 31
#>     msID site    site_short msNameKurz  site_long msKT  msOrt msOrtsteil msNABEL
#>    <int> <chr>   <chr>      <chr>       <chr>     <chr> <chr> <chr>        <int>
#>  1     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  2     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  3     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  4     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  5     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  6     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  7     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  8     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#>  9     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#> 10     1 Arb_St~ ARBS       Arbon Stad~ Arbon St~ TG    Arbon Stadthaus        2
#> # ... with 1,550 more rows, and 22 more variables: scDTV <chr>,
#> #   scDTVJahr <chr>, scLKWProzent <int>, scStrassenabstand <int>,
#> #   scEinwohner <int>, scEinwohnerJahr <chr>, scVerkehrslage <chr>,
#> #   scSiedlungsgroesse <chr>, scLagerelativSiedlung <chr>,
#> #   scIDBStandorttyp <chr>, scIDBStationstyp <chr>, scIDBBebauung <chr>,
#> #   scIDBVerkehr <chr>, x <int>, y <int>, masl <int>, parameter <chr>,
#> #   msmmBeginn <dttm>, msmmEnde <dttm>, source <chr>, ...
#> 

## get the cache as local storage
local <- store$get_local_storage()
local$list_chunks()
#> # A tibble: 20 x 7
#>    chunk_name   interval site  year  local.path   local.modification~ local.size
#>    <fs::path>   <chr>    <chr> <chr> <fs::path>   <dttm>              <fs::byte>
#>  1 d1/ZDHCu1pj~ d1       Zch_~ 2013  C:/Users/to~ 2021-12-04 21:16:52     19.84K
#>  2 d1/ZDHCu1pj~ d1       Zch_~ 2014  C:/Users/to~ 2021-12-04 21:16:52     76.76K
#>  3 d1/ZDHCu1pj~ d1       Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:52     86.04K
#>  4 h1/aDHCu1pj~ h1       Zch_~ 2013  C:/Users/to~ 2021-12-04 21:16:52    112.38K
#>  5 h1/aDHCu1pj~ h1       Zch_~ 2014  C:/Users/to~ 2021-12-04 21:16:52    447.54K
#>  6 h1/aDHCu1pj~ h1       Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:52    567.64K
#>  7 h8gl/aDhnbM~ h8gl     Zch_~ 2013  C:/Users/to~ 2021-12-04 21:16:52     23.56K
#>  8 h8gl/aDhnbM~ h8gl     Zch_~ 2014  C:/Users/to~ 2021-12-04 21:16:53     93.14K
#>  9 h8gl/aDhnbM~ h8gl     Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:53      92.2K
#> 10 m1/bTHCu1pj~ m1       Zch_~ 2013  C:/Users/to~ 2021-12-04 21:16:53      2.08K
#> 11 m1/bTHCu1pj~ m1       Zch_~ 2014  C:/Users/to~ 2021-12-04 21:16:53      5.64K
#> 12 m1/bTHCu1pj~ m1       Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:53      6.02K
#> 13 min30/bWluM~ min30    Zch_~ 2013  C:/Users/to~ 2021-12-04 21:16:53    203.55K
#> 14 min30/bWluM~ min30    Zch_~ 2014  C:/Users/to~ 2021-12-04 21:12:41    808.31K
#> 15 min30/bWluM~ min30    Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:53    924.68K
#> 16 min30/bWluM~ min30    Zch_~ 2014  C:/Users/to~ 2021-12-04 21:12:42      1.22M
#> 17 min30/bWluM~ min30    Zch_~ 2011  C:/Users/to~ 2021-12-04 21:12:41      1.21M
#> 18 min30/bWluM~ min30    Zch_~ 2012  C:/Users/to~ 2021-12-04 21:12:41      1.22M
#> 19 y1/eTHCu1pj~ y1       Zch_~ 2014  C:/Users/to~ 2021-12-04 21:16:53      1.49K
#> 20 y1/eTHCu1pj~ y1       Zch_~ 2015  C:/Users/to~ 2021-12-04 21:16:53      1.48K

## get cached data
local$get(site = "Zch_Stampfenbachstrasse", interval = "min30", year = 2011:2012)
#> # A tibble: 523,318 x 6
#>    starttime           site                    parameter interval unit   value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct>  <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    µg/m3  2.19 
#>  2 2011-01-01 00:00:00 Zch_Stampfenbachstrasse CO        min30    ppm    0.525
#>  3 2011-01-01 00:00:00 Zch_Stampfenbachstrasse Hr        min30    %Hr   91.9  
#>  4 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO2       min30    ppb   22.0  
#>  5 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO        min30    ppb   21.0  
#>  6 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NOx       min30    ppb   43.0  
#>  7 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10 
#>  8 2011-01-01 00:00:00 Zch_Stampfenbachstrasse PM10      min30    µg/m3 84.0  
#>  9 2011-01-01 00:00:00 Zch_Stampfenbachstrasse RainDur   min30    min    0    
#> 10 2011-01-01 00:00:00 Zch_Stampfenbachstrasse SO2       min30    ppb    1.90 
#> # ... with 523,308 more rows

## destroy store (careful removes all files on the disk)
store$read.only = FALSE
store$destroy("DELETE")
#> Cache for Store s3_example destroyed

## No examples for write operations