s3 storage

A s3 storage with flexible file format (default rds). The data format defines the data chunks per file. The data is cached locally. This local cache can be used as local storage. For performance enhancement it is recommended after acquiring all needed data from s3 to use the cache as local storage.

storage_s3_rds(name, format, bucket, prefix = NULL, region = NULL,
  read.only = TRUE)

Arguments

name
name of the store

format
data format of the store

bucket
name in aws s3

prefix
in aws s3

region
aws region

read.only
read only store. disables put, upload

Value

R6 class object of storage_s3

Fields

name

name of the store

format

data format of the store

bucket

s3 bucket containing the store

region

aws region of the bucket

prefix

of the s3 object keys. see Object Key and Metadata

path

local root of the store

data_path

local root of all chunks

data_s3

s3 root key of all chunks

content_path

local path to the rds file containing statistics of store content

content_s3

s3 object key to the rds file containing statistics of store content

meta_path

local root of all meta data files

meta_s3

s3 root key of all meta data files

read.only

flag for read.only usage of store. Default TRUE

ext

file extension for chunks. Default "rds"

read_function

function(file) for reading chunks from disk. Default base::readRDS()

write_function

function(object, file) for writing chunks to disk. Default base::saveRDS()

Methods

$get(filter=NULL, ...) get data from the store. The name of the arguments depend on the format. The filter argument is applied to each chunk.

$download(...) downloads data from s3 to the local cache. The dots arguments are used to filter the output of list_chunks(). Only the matching rows will be downloaded.

$put(data) puts the data into the store. Stops if store is read only

$upload() uploads content, meta data and all new and changed chunks to the s3 storage. For big additions to the store the recommend way is to use a local storage to modify the cache and use this function to apply the change.

$get_content() returns a tibble with the amount of data points per chunk per series

$list_chunks() get list of all chunks in s3 and local

$get_meta(key=NULL) get meta data. If key is omitted returns all the content of all files in a named list of tibbles, with the file name without extension as name. If key is supplied as argument only the list contains only the specified key.

$put_meta(...) puts meta data into the store. the name of the argument is used as file name and the value as data.

$get_local_storage() returns a storage to work with the cached data like a local storage

$destroy(confirmation) removes all files under path from the file system if "DELETE" is supplied as confirmation

Authentication

See the documentation of aws.signature for ways to provide the necessary informations. The simplest way is to use environment variables defined in a .Renviron file in the root directory of a RStudio Project:

AWS_ACCESS_KEY_ID = "xxxxxxxxxxxxxxxxxx"
AWS_SECRET_ACCESS_KEY = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
AWS_DEFAULT_REGION = "eu-central-1"

Examples

## init store, creates directory if necessary
format <- rOstluft::format_rolf()
store <- rOstluft::storage_s3_rds("s3_example", format, "rostluft", prefix = "aqmet")

## get all data min30 for 2011 and 2012
store$get(site = "Zch_Stampfenbachstrasse", interval = "min30", year = 2011:2012)

#> # A tibble: 488,322 x 6
#>    starttime           site                    parameter interval unit    value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct>   <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse CO        min30    ppm    0.525 
#>  2 2011-01-01 00:00:00 Zch_Stampfenbachstrasse Hr        min30    %Hr   91.9   
#>  3 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO2       min30    ppb   22.0   
#>  4 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO        min30    ppb   21.0   
#>  5 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NOx       min30    ppb   43.0   
#>  6 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10  
#>  7 2011-01-01 00:00:00 Zch_Stampfenbachstrasse PM10      min30    µg/m3 84.0   
#>  8 2011-01-01 00:00:00 Zch_Stampfenbachstrasse RainDur   min30    min    0     
#>  9 2011-01-01 00:00:00 Zch_Stampfenbachstrasse SO2       min30    ppb    1.90  
#> 10 2011-01-01 00:00:00 Zch_Stampfenbachstrasse StrGlo    min30    W/m2   0.0104
#> # ... with 488,312 more rows

## get only data for O3
store$get(year = 2011:2012, site = "Zch_Stampfenbachstrasse", interval = "min30",
          filter = parameter == "O3")

#> # A tibble: 34,996 x 6
#>    starttime           site                    parameter interval unit   value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct>  <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10 
#>  2 2011-01-01 00:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.05 
#>  3 2011-01-01 01:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.05 
#>  4 2011-01-01 01:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.00 
#>  5 2011-01-01 02:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.02 
#>  6 2011-01-01 02:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.05 
#>  7 2011-01-01 03:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    0.973
#>  8 2011-01-01 03:30:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.02 
#>  9 2011-01-01 04:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    3.10 
#> 10 2011-01-01 04:30:00 Zch_Stampfenbachstrasse O3        min30    ppb   18.7  
#> # ... with 34,986 more rows

## get NOx data from multiple stations
store$get(site = c("Zch_Stampfenbachstrasse", "Zch_Rosengartenstrasse"), interval = "min30",
          year = 2014, filter = parameter %in% c("NOx", "NO", "NO2"))

#> # A tibble: 104,547 x 6
#>    starttime           site                   parameter interval unit  value
#>    <dttm>              <fct>                  <fct>     <fct>    <fct> <dbl>
#>  1 2014-01-01 00:00:00 Zch_Rosengartenstrasse NO2       min30    ppb    18.6
#>  2 2014-01-01 00:00:00 Zch_Rosengartenstrasse NO        min30    ppb    42.8
#>  3 2014-01-01 00:00:00 Zch_Rosengartenstrasse NOx       min30    ppb    61.4
#>  4 2014-01-01 00:30:00 Zch_Rosengartenstrasse NO2       min30    ppb    21.3
#>  5 2014-01-01 00:30:00 Zch_Rosengartenstrasse NO        min30    ppb    55.9
#>  6 2014-01-01 00:30:00 Zch_Rosengartenstrasse NOx       min30    ppb    77.2
#>  7 2014-01-01 01:00:00 Zch_Rosengartenstrasse NO2       min30    ppb    28.2
#>  8 2014-01-01 01:00:00 Zch_Rosengartenstrasse NO        min30    ppb    88.1
#>  9 2014-01-01 01:00:00 Zch_Rosengartenstrasse NOx       min30    ppb   116. 
#> 10 2014-01-01 01:30:00 Zch_Rosengartenstrasse NO2       min30    ppb    39.8
#> # ... with 104,537 more rows

## get n data points grouped by intervall, station, parameter, year in the store
store$get_content()

#> # A tibble: 8,717 x 6
#>     year interval site             parameter unit                      n
#>    <dbl> <chr>    <fct>            <fct>     <fct>                 <int>
#>  1  2019 min10    ETHZ_Hönggerberg Hr        %                      9063
#>  2  2019 min10    ETHZ_Hönggerberg illum     logbase10ofvalueinLux  9063
#>  3  2019 min10    ETHZ_Hönggerberg p         hPa                    9064
#>  4  2019 min10    ETHZ_Hönggerberg RainSum   mm                     9063
#>  5  2019 min10    ETHZ_Hönggerberg T         °C                     9063
#>  6  2019 min10    ETHZ_Hönggerberg T_dew     °C                     9063
#>  7  2019 min10    ETHZ_Hönggerberg wcp_east  m/s                    9064
#>  8  2019 min10    ETHZ_Hönggerberg wcp_north m/s                    9064
#>  9  2019 min10    ETHZ_Hönggerberg WD        °                      9064
#> 10  2019 min10    ETHZ_Hönggerberg WD_max    °                      9064
#> # ... with 8,707 more rows

## get list of all chunks, show only local files
dplyr::filter(store$list_chunks(), !is.na(local.path))

#> # A tibble: 6 x 11
#>   chunk_name interval site  year  s3.key s3.lastmodified s3.etag s3.size
#>   <chr>      <chr>    <chr> <chr> <chr>  <chr>           <chr>   <fs::b>
#> 1 min30/bWl~ min30    Zch_~ 2013  aqmet~ 2019-03-05T16:~ 30b0e5~    168K
#> 2 min30/bWl~ min30    Zch_~ 2014  aqmet~ 2019-03-05T16:~ c2aa65~ 662.57K
#> 3 min30/bWl~ min30    Zch_~ 2015  aqmet~ 2019-03-05T16:~ 6c6620~ 780.05K
#> 4 min30/bWl~ min30    Zch_~ 2014  aqmet~ 2019-03-05T16:~ 638793~   1.07M
#> 5 min30/bWl~ min30    Zch_~ 2011  aqmet~ 2019-03-05T16:~ 934330~   1.07M
#> 6 min30/bWl~ min30    Zch_~ 2012  aqmet~ 2019-03-05T16:~ c78ad3~   1.07M
#> # ... with 3 more variables: local.path <fs::path>,
#> #   local.modification_time <dttm>, local.size <fs::bytes>

## download all data for site Zch_Rosengartenstrasse before 2016
store$download(site == "Zch_Rosengartenstrasse", year < 2016)

#> list()

## now there should be 2 more local files
dplyr::filter(store$list_chunks(), !is.na(local.path))

#> # A tibble: 6 x 11
#>   chunk_name interval site  year  s3.key s3.lastmodified s3.etag s3.size
#>   <chr>      <chr>    <chr> <chr> <chr>  <chr>           <chr>   <fs::b>
#> 1 min30/bWl~ min30    Zch_~ 2013  aqmet~ 2019-03-05T16:~ 30b0e5~    168K
#> 2 min30/bWl~ min30    Zch_~ 2014  aqmet~ 2019-03-05T16:~ c2aa65~ 662.57K
#> 3 min30/bWl~ min30    Zch_~ 2015  aqmet~ 2019-03-05T16:~ 6c6620~ 780.05K
#> 4 min30/bWl~ min30    Zch_~ 2014  aqmet~ 2019-03-05T16:~ 638793~   1.07M
#> 5 min30/bWl~ min30    Zch_~ 2011  aqmet~ 2019-03-05T16:~ 934330~   1.07M
#> 6 min30/bWl~ min30    Zch_~ 2012  aqmet~ 2019-03-05T16:~ c78ad3~   1.07M
#> # ... with 3 more variables: local.path <fs::path>,
#> #   local.modification_time <dttm>, local.size <fs::bytes>

## get all meta data
store$get_meta()

#> $meta
#> # A tibble: 1,202 x 12
#>    site  site_short site_long      x      y  masl parameter_origi~ parameter
#>    <fct> <fct>      <fct>      <dbl>  <dbl> <dbl> <fct>            <fct>    
#>  1 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 Hr               Hr       
#>  2 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 NO               NO       
#>  3 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 NO2              NO2      
#>  4 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 NOx              NOx      
#>  5 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 O3               O3       
#>  6 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 Ox               Ox       
#>  7 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 p                p        
#>  8 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 PM10             PM10     
#>  9 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 SO2              SO2      
#> 10 Arb_~ ARBS       Arbon St~ 2.75e6 1.26e6   407 StrGlo           StrGlo   
#> # ... with 1,192 more rows, and 4 more variables: unit <fct>, interval <fct>,
#> #   timezone_original <fct>, source <fct>
#> 
#> $meta_ecmwf
#> # A tibble: 24 x 12
#>    site_short site  site_long masl  parameter_origi~ unit  interval
#>    <chr>      <chr> <lgl>     <lgl> <chr>            <chr> <chr>   
#>  1 ZH_EraI    Züri~ NA        NA    t2m              °C    h1      
#>  2 ZH_EraI    Züri~ NA        NA    swvl1            m3/m3 h1      
#>  3 ZH_EraI    Züri~ NA        NA    sp               hPa   h1      
#>  4 ZH_EraI    Züri~ NA        NA    u10              m/s   h1      
#>  5 ZH_EraI    Züri~ NA        NA    v10              m/s   h1      
#>  6 ZH_EraI    Züri~ NA        NA    skt              °C    h1      
#>  7 ZH_EraI    Züri~ NA        NA    ishf             W/m2  h1      
#>  8 ZH_EraI    Züri~ NA        NA    tp               m     h1      
#>  9 ZH_EraI    Züri~ NA        NA    tco3             kg/m2 h1      
#> 10 ZH_EraI    Züri~ NA        NA    sund             s     h1      
#> # ... with 14 more rows, and 5 more variables: timezone_original <chr>,
#> #   source <chr>, x <dbl>, y <dbl>, parameter <chr>
#> 
#> $meta_ethz
#> # A tibble: 34 x 12
#>    site_short site_long site   masl parameter_origi~ unit  interval
#>    <fct>      <lgl>     <fct> <dbl> <fct>            <fct> <fct>   
#>  1 ETHZ_CHN   NA        ETHZ~   510 p_air            hPa   min10   
#>  2 ETHZ_CHN   NA        ETHZ~   510 T_air            °C    min10   
#>  3 ETHZ_CHN   NA        ETHZ~   510 T_dew            °C    min10   
#>  4 ETHZ_CHN   NA        ETHZ~   510 RH               %     min10   
#>  5 ETHZ_CHN   NA        ETHZ~   510 illum            logb~ min10   
#>  6 ETHZ_CHN   NA        ETHZ~   510 rain             mm    min10   
#>  7 ETHZ_CHN   NA        ETHZ~   510 wspd_vec         m/s   min10   
#>  8 ETHZ_CHN   NA        ETHZ~   510 wspd_scl         m/s   min10   
#>  9 ETHZ_CHN   NA        ETHZ~   510 wdir             °     min10   
#> 10 ETHZ_CHN   NA        ETHZ~   510 wspd_pk          m/s   min10   
#> # ... with 24 more rows, and 5 more variables: timezone_original <fct>,
#> #   source <fct>, x <dbl>, y <dbl>, parameter <fct>
#> 
#> $meta_hysplit
#> # A tibble: 2 x 12
#>   site_short site  site_long x     y     masl  parameter_origi~ parameter unit 
#>   <chr>      <chr> <chr>     <lgl> <lgl> <lgl> <chr>            <chr>     <lgl>
#> 1 ZH-Kasern~ ZH_K~ Zürich_K~ NA    NA    NA    hysplit_cluster  hysplit_~ NA   
#> 2 SG-BP-hys~ SG_B~ St. Gall~ NA    NA    NA    hysplit_cluster  hysplit_~ NA   
#> # ... with 3 more variables: interval <chr>, timezone_original <chr>,
#> #   source <chr>
#> 
#> $meta_meteoschweiz
#> # A tibble: 214 x 15
#>    site_short site  Länge Breite      x      y  masl parameter_origi~ source
#>    <chr>      <chr> <chr> <chr>   <dbl>  <dbl> <dbl> <chr>            <chr> 
#>  1 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 tre200s0         Meteo~
#>  2 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 ure200s0         Meteo~
#>  3 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 prestas0         Meteo~
#>  4 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 prestas0         Meteo~
#>  5 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 fkl010z1         Meteo~
#>  6 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 fkl010z0         Meteo~
#>  7 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 dkl010z0         Meteo~
#>  8 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 sre000z0         Meteo~
#>  9 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 sre000z0         Meteo~
#> 10 TAE        Aado~ 8°54' 47°29' 2.71e6 1.26e6   539 gre000z0         Meteo~
#> # ... with 204 more rows, and 6 more variables: unit <chr>, Beschreibung <chr>,
#> #   timezone_original <chr>, site_long <lgl>, interval <chr>, parameter <chr>
#> 
#> $meta_meteotest
#> # A tibble: 2 x 11
#>   site_short site_long site  lon   lat    masl parameter_origi~ unit  interval
#>   <chr>      <lgl>     <chr> <lgl> <lgl> <dbl> <chr>            <chr> <chr>   
#> 1 SMA_23m    NA        Züri~ NA    NA      581 T                °C    min10   
#> 2 ZUEberg    NA        Züri~ NA    NA      729 T                °C    min10   
#> # ... with 2 more variables: timezone_original <chr>, source <chr>
#> 
#> $meta_ostluft
#> # A tibble: 974 x 28
#>     msID site  site_short site_long msKT  scDTV scDTVJahr scLKWProzent
#>    <int> <chr> <chr>      <chr>     <chr> <chr> <chr>            <int>
#>  1     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  2     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  3     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  4     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  5     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  6     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  7     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  8     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  9     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#> 10     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#> # ... with 964 more rows, and 20 more variables: scStrassenabstand <int>,
#> #   scEinwohner <int>, scEinwohnerJahr <chr>, scVerkehrslage <chr>,
#> #   scSiedlungsgroesse <chr>, scLagerelativSiedlung <chr>,
#> #   scIDBStandorttyp <chr>, scIDBStationstyp <chr>, scIDBBebauung <chr>,
#> #   scIDBVerkehr <chr>, x <int>, y <int>, masl <int>, spBisDatum <dttm>,
#> #   parameter <chr>, msmmBeginn <dttm>, msmmEnde <dttm>, source <chr>,
#> #   parameter_original <chr>, timezone_original <chr>
#> 

## or a specific meta file
store$get_meta("meta_ostluft")

#> $meta_ostluft
#> # A tibble: 974 x 28
#>     msID site  site_short site_long msKT  scDTV scDTVJahr scLKWProzent
#>    <int> <chr> <chr>      <chr>     <chr> <chr> <chr>            <int>
#>  1     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  2     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  3     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  4     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  5     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  6     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  7     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  8     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#>  9     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#> 10     1 Arb_~ ARBS       Arbon St~ TG    7900  NA                   3
#> # ... with 964 more rows, and 20 more variables: scStrassenabstand <int>,
#> #   scEinwohner <int>, scEinwohnerJahr <chr>, scVerkehrslage <chr>,
#> #   scSiedlungsgroesse <chr>, scLagerelativSiedlung <chr>,
#> #   scIDBStandorttyp <chr>, scIDBStationstyp <chr>, scIDBBebauung <chr>,
#> #   scIDBVerkehr <chr>, x <int>, y <int>, masl <int>, spBisDatum <dttm>,
#> #   parameter <chr>, msmmBeginn <dttm>, msmmEnde <dttm>, source <chr>,
#> #   parameter_original <chr>, timezone_original <chr>
#> 

## get the cache as local storage
local <- store$get_local_storage()
local$list_chunks()

#> # A tibble: 6 x 7
#>   chunk_name interval site  year 
#>   <chr>      <chr>    <chr> <chr>
#> 1 min30/bWl~ min30    Zch_~ 2013 
#> 2 min30/bWl~ min30    Zch_~ 2014 
#> 3 min30/bWl~ min30    Zch_~ 2015 
#> 4 min30/bWl~ min30    Zch_~ 2014 
#> 5 min30/bWl~ min30    Zch_~ 2011 
#> 6 min30/bWl~ min30    Zch_~ 2012 
#> # ... with 3 more variables: local.path <fs::path>,
#> #   local.modification_time <dttm>, local.size <fs::bytes>

## get cached data
local$get(site = "Zch_Stampfenbachstrasse", interval = "min30", year = 2011:2012)

#> # A tibble: 488,322 x 6
#>    starttime           site                    parameter interval unit    value
#>    <dttm>              <fct>                   <fct>     <fct>    <fct>   <dbl>
#>  1 2011-01-01 00:00:00 Zch_Stampfenbachstrasse CO        min30    ppm    0.525 
#>  2 2011-01-01 00:00:00 Zch_Stampfenbachstrasse Hr        min30    %Hr   91.9   
#>  3 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO2       min30    ppb   22.0   
#>  4 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NO        min30    ppb   21.0   
#>  5 2011-01-01 00:00:00 Zch_Stampfenbachstrasse NOx       min30    ppb   43.0   
#>  6 2011-01-01 00:00:00 Zch_Stampfenbachstrasse O3        min30    ppb    1.10  
#>  7 2011-01-01 00:00:00 Zch_Stampfenbachstrasse PM10      min30    µg/m3 84.0   
#>  8 2011-01-01 00:00:00 Zch_Stampfenbachstrasse RainDur   min30    min    0     
#>  9 2011-01-01 00:00:00 Zch_Stampfenbachstrasse SO2       min30    ppb    1.90  
#> 10 2011-01-01 00:00:00 Zch_Stampfenbachstrasse StrGlo    min30    W/m2   0.0104
#> # ... with 488,312 more rows

## destroy store (careful removes all files on the disk)
store$destroy("DELETE")

#> Warning: Store still alive: read.only store or wrong confirmation phrase

## No examples for write operations