Reads a file from the SwissMetNet of MeteoSwiss containing multiple Exports

One peculiarity of MeteoSwiss is to provide multiple exports concatenated in one file, starting with a line containing a single space and seperated by two newlines. This function reads the whole file into the memory, splits the part and applies the read function to each part. Is the file too big for the available memory, the function split_smn() can split the file into multiple files each containing one part. On Linux the commandline tool csplit is another probably faster tool.

read_smn_multiple(fn, as_list = FALSE, encoding = "UTF-8", ...)

Arguments

fn
path to input file

as_list
return a list with each part as element rather than one concatenated tibble

encoding
encoding of the data file. Default = "UTF-8"

...
parameters passed to read_smn()

Value

tibble or list of tibbles in rOstluft long format structure

See also

Examples

fn <- system.file("extdata", "smn_multi.txt", package = "rOstluft.data", mustWork = TRUE)
read_smn_multiple(fn, as_list = TRUE)

#> [[1]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 ARH   dkl010z0  min10    NA    168  
#>  2 2018-01-01 01:00:00 ARH   dkl010z0  min10    NA    203  
#>  3 2018-01-01 01:10:00 ARH   dkl010z0  min10    NA    115  
#>  4 2018-01-01 01:20:00 ARH   dkl010z0  min10    NA     96  
#>  5 2018-01-01 01:30:00 ARH   dkl010z0  min10    NA     76  
#>  6 2018-01-01 01:40:00 ARH   dkl010z0  min10    NA     43  
#>  7 2018-01-01 00:50:00 ARH   fkl010z0  min10    NA      1.1
#>  8 2018-01-01 01:00:00 ARH   fkl010z0  min10    NA      0.6
#>  9 2018-01-01 01:10:00 ARH   fkl010z0  min10    NA      1.5
#> 10 2018-01-01 01:20:00 ARH   fkl010z0  min10    NA      1.9
#> # ... with 44 more rows
#> 
#> [[2]]
#> # A tibble: 18 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 BHF   fk1towz0  min10    NA      5.9
#>  2 2018-01-01 01:00:00 BHF   fk1towz0  min10    NA      7.2
#>  3 2018-01-01 01:10:00 BHF   fk1towz0  min10    NA      7.8
#>  4 2018-01-01 01:20:00 BHF   fk1towz0  min10    NA      9  
#>  5 2018-01-01 01:30:00 BHF   fk1towz0  min10    NA      9.7
#>  6 2018-01-01 01:40:00 BHF   fk1towz0  min10    NA      7.5
#>  7 2018-01-01 00:50:00 BHF   fkltowz1  min10    NA      8.5
#>  8 2018-01-01 01:00:00 BHF   fkltowz1  min10    NA     11  
#>  9 2018-01-01 01:10:00 BHF   fkltowz1  min10    NA     12  
#> 10 2018-01-01 01:20:00 BHF   fkltowz1  min10    NA     13  
#> 11 2018-01-01 01:30:00 BHF   fkltowz1  min10    NA     14.6
#> 12 2018-01-01 01:40:00 BHF   fkltowz1  min10    NA     15.7
#> 13 2018-01-01 00:50:00 BHF   ta1tows0  min10    NA      8.4
#> 14 2018-01-01 01:00:00 BHF   ta1tows0  min10    NA      8.8
#> 15 2018-01-01 01:10:00 BHF   ta1tows0  min10    NA      9  
#> 16 2018-01-01 01:20:00 BHF   ta1tows0  min10    NA      8.7
#> 17 2018-01-01 01:30:00 BHF   ta1tows0  min10    NA      8.2
#> 18 2018-01-01 01:40:00 BHF   ta1tows0  min10    NA      7.2
#> 
#> [[3]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 CHU   dkl010z0  min10    NA    216  
#>  2 2018-01-01 01:00:00 CHU   dkl010z0  min10    NA    209  
#>  3 2018-01-01 01:10:00 CHU   dkl010z0  min10    NA    216  
#>  4 2018-01-01 01:20:00 CHU   dkl010z0  min10    NA    227  
#>  5 2018-01-01 01:30:00 CHU   dkl010z0  min10    NA    212  
#>  6 2018-01-01 01:40:00 CHU   dkl010z0  min10    NA    200  
#>  7 2018-01-01 00:50:00 CHU   fkl010z0  min10    NA      1.7
#>  8 2018-01-01 01:00:00 CHU   fkl010z0  min10    NA      1.9
#>  9 2018-01-01 01:10:00 CHU   fkl010z0  min10    NA      2  
#> 10 2018-01-01 01:20:00 CHU   fkl010z0  min10    NA      2.6
#> # ... with 44 more rows
#> 
#> [[4]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 EBK   dkl010z0  min10    NA    179  
#>  2 2018-01-01 01:00:00 EBK   dkl010z0  min10    NA    161  
#>  3 2018-01-01 01:10:00 EBK   dkl010z0  min10    NA    178  
#>  4 2018-01-01 01:20:00 EBK   dkl010z0  min10    NA    114  
#>  5 2018-01-01 01:30:00 EBK   dkl010z0  min10    NA    135  
#>  6 2018-01-01 01:40:00 EBK   dkl010z0  min10    NA     66  
#>  7 2018-01-01 00:50:00 EBK   fkl010z0  min10    NA      0.9
#>  8 2018-01-01 01:00:00 EBK   fkl010z0  min10    NA      0.9
#>  9 2018-01-01 01:10:00 EBK   fkl010z0  min10    NA      0.9
#> 10 2018-01-01 01:20:00 EBK   fkl010z0  min10    NA      1  
#> # ... with 44 more rows
#> 
#> [[5]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 GLA   dkl010z0  min10    NA    140  
#>  2 2018-01-01 01:00:00 GLA   dkl010z0  min10    NA    105  
#>  3 2018-01-01 01:10:00 GLA   dkl010z0  min10    NA    246  
#>  4 2018-01-01 01:20:00 GLA   dkl010z0  min10    NA    100  
#>  5 2018-01-01 01:30:00 GLA   dkl010z0  min10    NA      0  
#>  6 2018-01-01 01:40:00 GLA   dkl010z0  min10    NA    228  
#>  7 2018-01-01 00:50:00 GLA   fkl010z0  min10    NA      1.3
#>  8 2018-01-01 01:00:00 GLA   fkl010z0  min10    NA      0.5
#>  9 2018-01-01 01:10:00 GLA   fkl010z0  min10    NA      1.1
#> 10 2018-01-01 01:20:00 GLA   fkl010z0  min10    NA      1.1
#> # ... with 44 more rows
#> 
#> [[6]]
#> # A tibble: 18 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 GUB   fk1towz0  min10    NA      6.1
#>  2 2018-01-01 01:00:00 GUB   fk1towz0  min10    NA      7.5
#>  3 2018-01-01 01:10:00 GUB   fk1towz0  min10    NA      6.6
#>  4 2018-01-01 01:20:00 GUB   fk1towz0  min10    NA      5.3
#>  5 2018-01-01 01:30:00 GUB   fk1towz0  min10    NA      3.7
#>  6 2018-01-01 01:40:00 GUB   fk1towz0  min10    NA      4.9
#>  7 2018-01-01 00:50:00 GUB   fkltowz1  min10    NA     12.9
#>  8 2018-01-01 01:00:00 GUB   fkltowz1  min10    NA     14.4
#>  9 2018-01-01 01:10:00 GUB   fkltowz1  min10    NA     16.3
#> 10 2018-01-01 01:20:00 GUB   fkltowz1  min10    NA     11.1
#> 11 2018-01-01 01:30:00 GUB   fkltowz1  min10    NA      7.9
#> 12 2018-01-01 01:40:00 GUB   fkltowz1  min10    NA      9.6
#> 13 2018-01-01 00:50:00 GUB   ta1tows0  min10    NA      7.5
#> 14 2018-01-01 01:00:00 GUB   ta1tows0  min10    NA      8  
#> 15 2018-01-01 01:10:00 GUB   ta1tows0  min10    NA      7.8
#> 16 2018-01-01 01:20:00 GUB   ta1tows0  min10    NA      7.8
#> 17 2018-01-01 01:30:00 GUB   ta1tows0  min10    NA      7.4
#> 18 2018-01-01 01:40:00 GUB   ta1tows0  min10    NA      6.8
#> 
#> [[7]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 HOE   dkl010z0  min10    NA    251  
#>  2 2018-01-01 01:00:00 HOE   dkl010z0  min10    NA    270  
#>  3 2018-01-01 01:10:00 HOE   dkl010z0  min10    NA    271  
#>  4 2018-01-01 01:20:00 HOE   dkl010z0  min10    NA    274  
#>  5 2018-01-01 01:30:00 HOE   dkl010z0  min10    NA    278  
#>  6 2018-01-01 01:40:00 HOE   dkl010z0  min10    NA    280  
#>  7 2018-01-01 00:50:00 HOE   fkl010z0  min10    NA     12.3
#>  8 2018-01-01 01:00:00 HOE   fkl010z0  min10    NA      9.3
#>  9 2018-01-01 01:10:00 HOE   fkl010z0  min10    NA     11.7
#> 10 2018-01-01 01:20:00 HOE   fkl010z0  min10    NA     11.2
#> # ... with 44 more rows
#> 
#> [[8]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 KLO   dkl010z0  min10    NA    224  
#>  2 2018-01-01 01:00:00 KLO   dkl010z0  min10    NA    231  
#>  3 2018-01-01 01:10:00 KLO   dkl010z0  min10    NA    237  
#>  4 2018-01-01 01:20:00 KLO   dkl010z0  min10    NA    244  
#>  5 2018-01-01 01:30:00 KLO   dkl010z0  min10    NA    260  
#>  6 2018-01-01 01:40:00 KLO   dkl010z0  min10    NA    271  
#>  7 2018-01-01 00:50:00 KLO   fkl010z0  min10    NA      5.7
#>  8 2018-01-01 01:00:00 KLO   fkl010z0  min10    NA      5.9
#>  9 2018-01-01 01:10:00 KLO   fkl010z0  min10    NA      7.1
#> 10 2018-01-01 01:20:00 KLO   fkl010z0  min10    NA      6.6
#> # ... with 44 more rows
#> 
#> [[9]]
#> # A tibble: 42 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 LAE   dkl010z0  min10    NA    208  
#>  2 2018-01-01 01:00:00 LAE   dkl010z0  min10    NA    214  
#>  3 2018-01-01 01:10:00 LAE   dkl010z0  min10    NA    220  
#>  4 2018-01-01 01:20:00 LAE   dkl010z0  min10    NA    252  
#>  5 2018-01-01 01:30:00 LAE   dkl010z0  min10    NA    277  
#>  6 2018-01-01 01:40:00 LAE   dkl010z0  min10    NA    287  
#>  7 2018-01-01 00:50:00 LAE   fkl010z1  min10    NA     12  
#>  8 2018-01-01 01:00:00 LAE   fkl010z1  min10    NA     14.3
#>  9 2018-01-01 01:10:00 LAE   fkl010z1  min10    NA     13.2
#> 10 2018-01-01 01:20:00 LAE   fkl010z1  min10    NA     12.8
#> # ... with 32 more rows
#> 
#> [[10]]
#> # A tibble: 18 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 LCK   fk1towz0  min10    NA      7.7
#>  2 2018-01-01 01:00:00 LCK   fk1towz0  min10    NA      8.5
#>  3 2018-01-01 01:10:00 LCK   fk1towz0  min10    NA     10.5
#>  4 2018-01-01 01:20:00 LCK   fk1towz0  min10    NA      9.9
#>  5 2018-01-01 01:30:00 LCK   fk1towz0  min10    NA      8.2
#>  6 2018-01-01 01:40:00 LCK   fk1towz0  min10    NA      6.7
#>  7 2018-01-01 00:50:00 LCK   fkltowz1  min10    NA     10.7
#>  8 2018-01-01 01:00:00 LCK   fkltowz1  min10    NA     12.5
#>  9 2018-01-01 01:10:00 LCK   fkltowz1  min10    NA     15.5
#> 10 2018-01-01 01:20:00 LCK   fkltowz1  min10    NA     13.2
#> 11 2018-01-01 01:30:00 LCK   fkltowz1  min10    NA     11.1
#> 12 2018-01-01 01:40:00 LCK   fkltowz1  min10    NA      9.6
#> 13 2018-01-01 00:50:00 LCK   ta1tows0  min10    NA      7.3
#> 14 2018-01-01 01:00:00 LCK   ta1tows0  min10    NA      7.4
#> 15 2018-01-01 01:10:00 LCK   ta1tows0  min10    NA      7.6
#> 16 2018-01-01 01:20:00 LCK   ta1tows0  min10    NA      7.7
#> 17 2018-01-01 01:30:00 LCK   ta1tows0  min10    NA      7.7
#> 18 2018-01-01 01:40:00 LCK   ta1tows0  min10    NA      7.5
#> 
#> [[11]]
#> # A tibble: 12 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 NEE   fk1towz0  min10    NA      7.9
#>  2 2018-01-01 01:00:00 NEE   fk1towz0  min10    NA      8  
#>  3 2018-01-01 01:10:00 NEE   fk1towz0  min10    NA      6.6
#>  4 2018-01-01 01:20:00 NEE   fk1towz0  min10    NA      5  
#>  5 2018-01-01 01:30:00 NEE   fk1towz0  min10    NA      3.6
#>  6 2018-01-01 01:40:00 NEE   fk1towz0  min10    NA      3.2
#>  7 2018-01-01 00:50:00 NEE   fkltowz1  min10    NA     12.7
#>  8 2018-01-01 01:00:00 NEE   fkltowz1  min10    NA     12.1
#>  9 2018-01-01 01:10:00 NEE   fkltowz1  min10    NA     10.3
#> 10 2018-01-01 01:20:00 NEE   fkltowz1  min10    NA     11.1
#> 11 2018-01-01 01:30:00 NEE   fkltowz1  min10    NA      7.5
#> 12 2018-01-01 01:40:00 NEE   fkltowz1  min10    NA      6.6
#> 
#> [[12]]
#> # A tibble: 42 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 OBR   dkl010z0  min10    NA    259  
#>  2 2018-01-01 01:00:00 OBR   dkl010z0  min10    NA    248  
#>  3 2018-01-01 01:10:00 OBR   dkl010z0  min10    NA    248  
#>  4 2018-01-01 01:20:00 OBR   dkl010z0  min10    NA    225  
#>  5 2018-01-01 01:30:00 OBR   dkl010z0  min10    NA    265  
#>  6 2018-01-01 01:40:00 OBR   dkl010z0  min10    NA    288  
#>  7 2018-01-01 00:50:00 OBR   fkl010z0  min10    NA      3.1
#>  8 2018-01-01 01:00:00 OBR   fkl010z0  min10    NA      2.6
#>  9 2018-01-01 01:10:00 OBR   fkl010z0  min10    NA      1.9
#> 10 2018-01-01 01:20:00 OBR   fkl010z0  min10    NA      2.4
#> # ... with 32 more rows
#> 
#> [[13]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 PIL   dkl010z0  min10    NA    212  
#>  2 2018-01-01 01:00:00 PIL   dkl010z0  min10    NA    213  
#>  3 2018-01-01 01:10:00 PIL   dkl010z0  min10    NA    210  
#>  4 2018-01-01 01:20:00 PIL   dkl010z0  min10    NA    164  
#>  5 2018-01-01 01:30:00 PIL   dkl010z0  min10    NA    195  
#>  6 2018-01-01 01:40:00 PIL   dkl010z0  min10    NA    205  
#>  7 2018-01-01 00:50:00 PIL   fkl010z0  min10    NA      8.5
#>  8 2018-01-01 01:00:00 PIL   fkl010z0  min10    NA      8.1
#>  9 2018-01-01 01:10:00 PIL   fkl010z0  min10    NA      4.6
#> 10 2018-01-01 01:20:00 PIL   fkl010z0  min10    NA      1.9
#> # ... with 44 more rows
#> 
#> [[14]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 SAE   dkl010z0  min10    NA    256  
#>  2 2018-01-01 01:00:00 SAE   dkl010z0  min10    NA    258  
#>  3 2018-01-01 01:10:00 SAE   dkl010z0  min10    NA    255  
#>  4 2018-01-01 01:20:00 SAE   dkl010z0  min10    NA    257  
#>  5 2018-01-01 01:30:00 SAE   dkl010z0  min10    NA    253  
#>  6 2018-01-01 01:40:00 SAE   dkl010z0  min10    NA    253  
#>  7 2018-01-01 00:50:00 SAE   fkl010z0  min10    NA     13.3
#>  8 2018-01-01 01:00:00 SAE   fkl010z0  min10    NA     13.9
#>  9 2018-01-01 01:10:00 SAE   fkl010z0  min10    NA     14.8
#> 10 2018-01-01 01:20:00 SAE   fkl010z0  min10    NA     15.9
#> # ... with 44 more rows
#> 
#> [[15]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 SHA   dkl010z0  min10    NA    253  
#>  2 2018-01-01 01:00:00 SHA   dkl010z0  min10    NA    255  
#>  3 2018-01-01 01:10:00 SHA   dkl010z0  min10    NA    260  
#>  4 2018-01-01 01:20:00 SHA   dkl010z0  min10    NA    257  
#>  5 2018-01-01 01:30:00 SHA   dkl010z0  min10    NA    254  
#>  6 2018-01-01 01:40:00 SHA   dkl010z0  min10    NA    255  
#>  7 2018-01-01 00:50:00 SHA   fkl010z0  min10    NA      6  
#>  8 2018-01-01 01:00:00 SHA   fkl010z0  min10    NA      6.1
#>  9 2018-01-01 01:10:00 SHA   fkl010z0  min10    NA      6.4
#> 10 2018-01-01 01:20:00 SHA   fkl010z0  min10    NA      6.6
#> # ... with 44 more rows
#> 
#> [[16]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 SMA   dkl010z0  min10    NA    248  
#>  2 2018-01-01 01:00:00 SMA   dkl010z0  min10    NA    244  
#>  3 2018-01-01 01:10:00 SMA   dkl010z0  min10    NA    241  
#>  4 2018-01-01 01:20:00 SMA   dkl010z0  min10    NA    237  
#>  5 2018-01-01 01:30:00 SMA   dkl010z0  min10    NA    237  
#>  6 2018-01-01 01:40:00 SMA   dkl010z0  min10    NA    240  
#>  7 2018-01-01 00:50:00 SMA   fkl010z0  min10    NA      7.2
#>  8 2018-01-01 01:00:00 SMA   fkl010z0  min10    NA      6.5
#>  9 2018-01-01 01:10:00 SMA   fkl010z0  min10    NA      8.2
#> 10 2018-01-01 01:20:00 SMA   fkl010z0  min10    NA      8.9
#> # ... with 44 more rows
#> 
#> [[17]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 STG   dkl010z0  min10    NA    220  
#>  2 2018-01-01 01:00:00 STG   dkl010z0  min10    NA    218  
#>  3 2018-01-01 01:10:00 STG   dkl010z0  min10    NA    228  
#>  4 2018-01-01 01:20:00 STG   dkl010z0  min10    NA    227  
#>  5 2018-01-01 01:30:00 STG   dkl010z0  min10    NA    228  
#>  6 2018-01-01 01:40:00 STG   dkl010z0  min10    NA    227  
#>  7 2018-01-01 00:50:00 STG   fkl010z0  min10    NA      5.5
#>  8 2018-01-01 01:00:00 STG   fkl010z0  min10    NA      4.8
#>  9 2018-01-01 01:10:00 STG   fkl010z0  min10    NA      5.3
#> 10 2018-01-01 01:20:00 STG   fkl010z0  min10    NA      5.2
#> # ... with 44 more rows
#> 
#> [[18]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 TAE   dkl010z0  min10    NA    232  
#>  2 2018-01-01 01:00:00 TAE   dkl010z0  min10    NA    233  
#>  3 2018-01-01 01:10:00 TAE   dkl010z0  min10    NA    265  
#>  4 2018-01-01 01:20:00 TAE   dkl010z0  min10    NA    241  
#>  5 2018-01-01 01:30:00 TAE   dkl010z0  min10    NA    232  
#>  6 2018-01-01 01:40:00 TAE   dkl010z0  min10    NA    239  
#>  7 2018-01-01 00:50:00 TAE   fkl010z0  min10    NA      1.5
#>  8 2018-01-01 01:00:00 TAE   fkl010z0  min10    NA      1.2
#>  9 2018-01-01 01:10:00 TAE   fkl010z0  min10    NA      2.7
#> 10 2018-01-01 01:20:00 TAE   fkl010z0  min10    NA      4.1
#> # ... with 44 more rows
#> 
#> [[19]]
#> # A tibble: 30 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 UEB   fk1towz0  min10    NA     17.3
#>  2 2018-01-01 01:00:00 UEB   fk1towz0  min10    NA     17.1
#>  3 2018-01-01 01:10:00 UEB   fk1towz0  min10    NA     18.4
#>  4 2018-01-01 01:20:00 UEB   fk1towz0  min10    NA     18.5
#>  5 2018-01-01 01:30:00 UEB   fk1towz0  min10    NA     16.1
#>  6 2018-01-01 01:40:00 UEB   fk1towz0  min10    NA     13.5
#>  7 2018-01-01 00:50:00 UEB   fkltowz1  min10    NA     18.8
#>  8 2018-01-01 01:00:00 UEB   fkltowz1  min10    NA     22  
#>  9 2018-01-01 01:10:00 UEB   fkltowz1  min10    NA     22.6
#> 10 2018-01-01 01:20:00 UEB   fkltowz1  min10    NA     26.3
#> # ... with 20 more rows
#> 
#> [[20]]
#> # A tibble: 6 x 6
#>   starttime           site  parameter interval unit  value
#>   <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#> 1 2018-01-01 00:00:00 UEB   uretowhs  h1       NA     85.1
#> 2 2018-01-01 01:00:00 UEB   uretowhs  h1       NA     94.5
#> 3 2018-01-01 02:00:00 UEB   uretowhs  h1       NA    100  
#> 4 2018-01-01 03:00:00 UEB   uretowhs  h1       NA     97.5
#> 5 2018-01-01 04:00:00 UEB   uretowhs  h1       NA     88  
#> 6 2018-01-01 05:00:00 UEB   uretowhs  h1       NA     82.4
#> 
#> [[21]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2019-01-31 00:50:00 TAE   tre200s0  min10    °C     -5.8
#>  2 2019-01-31 01:00:00 TAE   tre200s0  min10    °C     -5.7
#>  3 2019-01-31 01:10:00 TAE   tre200s0  min10    °C     -6.6
#>  4 2019-01-31 01:20:00 TAE   tre200s0  min10    °C     -5.7
#>  5 2019-01-31 01:30:00 TAE   tre200s0  min10    °C     -7.3
#>  6 2019-01-31 01:40:00 TAE   tre200s0  min10    °C     -6.7
#>  7 2019-01-31 00:50:00 TAE   ure200s0  min10    %      81  
#>  8 2019-01-31 01:00:00 TAE   ure200s0  min10    %      78.2
#>  9 2019-01-31 01:10:00 TAE   ure200s0  min10    %      81.8
#> 10 2019-01-31 01:20:00 TAE   ure200s0  min10    %      78.4
#> # ... with 44 more rows
#> 
#> [[22]]
#> # A tibble: 31 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2019-01-31 00:50:00 UEB   ta1tows0  min10    °C     -3.3
#>  2 2019-01-31 01:00:00 UEB   ta1tows0  min10    °C     -3.3
#>  3 2019-01-31 01:10:00 UEB   ta1tows0  min10    °C     -3.3
#>  4 2019-01-31 01:20:00 UEB   ta1tows0  min10    °C     -3.3
#>  5 2019-01-31 01:30:00 UEB   ta1tows0  min10    °C     -3.3
#>  6 2019-01-31 01:40:00 UEB   ta1tows0  min10    °C     -3.3
#>  7 2019-01-31 00:50:00 UEB   uretowhs  min10    %      47.6
#>  8 2019-01-31 00:50:00 UEB   fk1towz0  min10    m/s     4.7
#>  9 2019-01-31 01:00:00 UEB   fk1towz0  min10    m/s     4.4
#> 10 2019-01-31 01:10:00 UEB   fk1towz0  min10    m/s     3.7
#> # ... with 21 more rows
#> 
#> [[23]]
#> # A tibble: 54 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2018-01-01 00:50:00 VAD   dkl010z0  min10    NA    138  
#>  2 2018-01-01 01:00:00 VAD   dkl010z0  min10    NA    137  
#>  3 2018-01-01 01:10:00 VAD   dkl010z0  min10    NA    129  
#>  4 2018-01-01 01:20:00 VAD   dkl010z0  min10    NA    136  
#>  5 2018-01-01 01:30:00 VAD   dkl010z0  min10    NA    130  
#>  6 2018-01-01 01:40:00 VAD   dkl010z0  min10    NA    133  
#>  7 2018-01-01 00:50:00 VAD   fkl010z0  min10    NA      2.7
#>  8 2018-01-01 01:00:00 VAD   fkl010z0  min10    NA      3.8
#>  9 2018-01-01 01:10:00 VAD   fkl010z0  min10    NA      3.5
#> 10 2018-01-01 01:20:00 VAD   fkl010z0  min10    NA      3.4
#> # ... with 44 more rows
#> 
#> [[24]]
#> # A tibble: 24 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2010-01-01 00:00:00 KLO   dkl010b0  h1       NA    319  
#>  2 2010-01-01 01:00:00 KLO   dkl010b0  h1       NA    317  
#>  3 2010-01-01 02:00:00 KLO   dkl010b0  h1       NA    360  
#>  4 2010-01-01 03:00:00 KLO   dkl010b0  h1       NA    117  
#>  5 2010-01-01 04:00:00 KLO   dkl010b0  h1       NA    360  
#>  6 2010-01-01 05:00:00 KLO   dkl010b0  h1       NA    321  
#>  7 2010-01-01 00:00:00 KLO   fkl010b0  h1       NA      1.5
#>  8 2010-01-01 01:00:00 KLO   fkl010b0  h1       NA      1.7
#>  9 2010-01-01 02:00:00 KLO   fkl010b0  h1       NA      0.7
#> 10 2010-01-01 03:00:00 KLO   fkl010b0  h1       NA      1.4
#> # ... with 14 more rows
#> 
#> [[25]]
#> # A tibble: 18 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2010-01-01 00:00:00 UEB   dk1towb0  h1       NA    249  
#>  2 2010-01-01 01:00:00 UEB   dk1towb0  h1       NA    259  
#>  3 2010-01-01 02:00:00 UEB   dk1towb0  h1       NA    237  
#>  4 2010-01-01 03:00:00 UEB   dk1towb0  h1       NA    225  
#>  5 2010-01-01 04:00:00 UEB   dk1towb0  h1       NA    153  
#>  6 2010-01-01 05:00:00 UEB   dk1towb0  h1       NA    129  
#>  7 2010-01-01 00:00:00 UEB   fk1towb0  h1       NA      1.7
#>  8 2010-01-01 01:00:00 UEB   fk1towb0  h1       NA      2.1
#>  9 2010-01-01 02:00:00 UEB   fk1towb0  h1       NA      1.8
#> 10 2010-01-01 03:00:00 UEB   fk1towb0  h1       NA      0.3
#> 11 2010-01-01 04:00:00 UEB   fk1towb0  h1       NA      0.6
#> 12 2010-01-01 05:00:00 UEB   fk1towb0  h1       NA      1.9
#> 13 2010-01-01 00:00:00 UEB   ta1towb0  h1       NA      3.7
#> 14 2010-01-01 01:00:00 UEB   ta1towb0  h1       NA      3.5
#> 15 2010-01-01 02:00:00 UEB   ta1towb0  h1       NA      3.2
#> 16 2010-01-01 03:00:00 UEB   ta1towb0  h1       NA      3.1
#> 17 2010-01-01 04:00:00 UEB   ta1towb0  h1       NA      3.3
#> 18 2010-01-01 05:00:00 UEB   ta1towb0  h1       NA      3.3
#> 
#> [[26]]
#> # A tibble: 55 x 6
#>    starttime           site  parameter interval unit  value
#>    <dttm>              <fct> <fct>     <fct>    <fct> <dbl>
#>  1 2017-12-31 01:00:00 RHW   wkcap1d0  d1       NA        2
#>  2 2018-01-01 01:00:00 RHW   wkcap1d0  d1       NA        1
#>  3 2018-01-02 01:00:00 RHW   wkcap1d0  d1       NA        7
#>  4 2018-01-03 01:00:00 RHW   wkcap1d0  d1       NA        7
#>  5 2018-01-04 01:00:00 RHW   wkcap1d0  d1       NA        7
#>  6 2017-12-31 01:00:00 RHW   wkcap2d0  d1       NA       11
#>  7 2018-01-01 01:00:00 RHW   wkcap2d0  d1       NA        1
#>  8 2018-01-02 01:00:00 RHW   wkcap2d0  d1       NA       14
#>  9 2018-01-03 01:00:00 RHW   wkcap2d0  d1       NA       14
#> 10 2018-01-04 01:00:00 RHW   wkcap2d0  d1       NA       14
#> # ... with 45 more rows
#>